+++ packages/sdk/src/homa/utils/get-estimate-redeem-result.ts
@@ -16,7 +16,7 @@ export function getEstimateRedeemResult(
       request: amount,
       receive: redeemStaking,
       fee,
-      canTryFastReddem: env.toBondPool.gt(redeemStaking),
+      canTryFastRedeem: env.toBondPool.gt(redeemStaking),
       env
     };
   }
@@ -25,7 +25,7 @@ export function getEstimateRedeemResult(
     request: amount,
     receive: convertLiquidToStaking(env.exchangeRate, amount),
     fee: FixedPointNumber.ZERO,
-    canTryFastReddem: false,
+    canTryFastRedeem: false,
     env
   };
 }