+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # CHANGELOG
 
+## 4.0.2-18 Mar 02, 2022
+Changes:
+  - fix error spell, canTryFastReddem -> canTryFastRedeem
+  - support query LiquidCrowdloanCurrency Price util
+  - homa sdk support promise
+
 ## 4.0.2-9 Oct 01, 2022
 Change:
   - Bump polkadot.js version to 7.2.1
   - Support LCDOT price
+
 ## 3.0.2 Oct 29, 2021
 Changes:
   - Add StableAsset types